fix(db): CockroachDB float/int 나누기 오류 및 시퀀스 동기화
1. ::float / int → ::float8 / ::float8 전면 수정
   (assessment.py, winner_dist.py, statistics.py)
   CockroachDB에서 <float> / <int> 연산자 미지원으로
   get_quick_decision → get_assessment_rate_dist 전 경로 실패하던 문제 해결.

2. prediction_logs_v2 insert 실패 시 db.rollback() 추가 (decision_service.py)
   UniqueViolation 이후 세션이 오염되어 후속 쿼리가 전부 실패하던 문제 해결.

3. startup에 CockroachDB 시퀀스 동기화 추가 (main.py)
   데이터 import 후 시퀀스가 1로 리셋되어 duplicate key 오류 반복 발생 방지.

+++ bid-system/backend/app/ml/assessment.py
@@ -873,8 +873,8 @@ def predict_a_ratio(
     if agency_id and industry_id:
         row = db.execute(text("""
             SELECT
-                AVG(ib.estimated_amount::float / ib.base_amount) AS avg_r,
-                STDDEV(ib.estimated_amount::float / ib.base_amount) AS std_r,
+                AVG(ib.estimated_amount::float8 / ib.base_amount::float8) AS avg_r,
+                STDDEV(ib.estimated_amount::float8 / ib.base_amount::float8) AS std_r,
                 COUNT(*) AS cnt
             FROM inpo21c_bids ib
             JOIN bids b ON b.announcement_no = ib.announcement_no
@@ -884,7 +884,7 @@ def predict_a_ratio(
             )
             WHERE a.id = :ag AND b.industry_id = :ind
               AND ib.estimated_amount IS NOT NULL AND ib.base_amount > 0
-              AND ib.estimated_amount::float / ib.base_amount BETWEEN 0.70 AND 1.10
+              AND ib.estimated_amount::float8 / ib.base_amount::float8 BETWEEN 0.70 AND 1.10
         """), {"ag": agency_id, "ind": industry_id}).fetchone()
         if row and row[2] and int(row[2]) >= MIN_SAMPLES and 0.70 < float(row[0]) < 1.10:
             return {
@@ -898,8 +898,8 @@ def predict_a_ratio(
     if agency_id:
         row = db.execute(text("""
             SELECT
-                AVG(ib.estimated_amount::float / ib.base_amount),
-                STDDEV(ib.estimated_amount::float / ib.base_amount),
+                AVG(ib.estimated_amount::float8 / ib.base_amount::float8),
+                STDDEV(ib.estimated_amount::float8 / ib.base_amount::float8),
                 COUNT(*)
             FROM inpo21c_bids ib
             JOIN agencies a ON (
@@ -909,7 +909,7 @@ def predict_a_ratio(
             )
             WHERE a.id = :ag
               AND ib.estimated_amount IS NOT NULL AND ib.base_amount > 0
-              AND ib.estimated_amount::float / ib.base_amount BETWEEN 0.70 AND 1.10
+              AND ib.estimated_amount::float8 / ib.base_amount::float8 BETWEEN 0.70 AND 1.10
         """), {"ag": agency_id}).fetchone()
         if row and row[2] and int(row[2]) >= MIN_SAMPLES and 0.70 < float(row[0]) < 1.10:
             return {
